+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/templating/mail/BaseRedisTest.java
@@ -3,6 +3,8 @@ package it.ozimov.springboot.templating.mail;
 import it.ozimov.springboot.templating.mail.model.EmailSchedulingData;
 import it.ozimov.springboot.templating.mail.service.EmailEmbeddedRedis;
 import lombok.NonNull;
+import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Before;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,9 +22,20 @@ import org.springframework.data.redis.serializer.JdkSerializationRedisSerializer
 import org.springframework.data.redis.serializer.RedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 import org.springframework.jdbc.datasource.DataSourceTransactionManager;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.TestContext;
+import org.springframework.test.context.TestExecutionListener;
+import org.springframework.test.context.TestExecutionListeners;
+import org.springframework.test.context.support.AbstractTestExecutionListener;
+import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
+import org.springframework.test.context.support.DirtiesContextTestExecutionListener;
 import org.springframework.test.context.transaction.AfterTransaction;
 import org.springframework.test.context.transaction.BeforeTransaction;
 import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionDefinition;
+import org.springframework.transaction.TransactionException;
+import org.springframework.transaction.support.AbstractPlatformTransactionManager;
+import org.springframework.transaction.support.DefaultTransactionStatus;
 import org.springframework.util.SocketUtils;
 import redis.clients.jedis.JedisShardInfo;
 import redis.embedded.RedisServer;
@@ -35,9 +48,13 @@ import java.sql.SQLException;
 import java.util.Objects;
 
 import static java.util.Objects.nonNull;
+import static org.mockito.Mockito.*;
 import static org.mockito.Mockito.mock;
 
-public abstract class BaseRedisTest implements ContextBasedTest {
+@TestExecutionListeners(
+        mergeMode =TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS,
+        listeners = {BaseRedisTest.class})
+public abstract class BaseRedisTest  implements ContextBasedTest, TestExecutionListener {
 
     protected AfterTransactionAssertion afterTransactionAssertion;
     protected BeforeTransactionAssertion beforeTransactionAssertion;
@@ -57,6 +74,9 @@ public abstract class BaseRedisTest implements ContextBasedTest {
     protected void setBeforeTransactionAssertion(@NonNull final BeforeTransactionAssertion beforeTransactionAssertion) {
         this.beforeTransactionAssertion = beforeTransactionAssertion;
     }
+
+    @Autowired
+    private RedisServer redisServer;
 
     @Autowired
     private RedisConnectionFactory connectionFactory;
@@ -81,6 +101,7 @@ public abstract class BaseRedisTest implements ContextBasedTest {
         if (nonNull(beforeTransactionAssertion)) {
             final RedisConnection connection = connectionFactory.getConnection();
             beforeTransactionAssertion.assertBeforeTransaction(connection);
+            connection.close();
         }
     }
 
@@ -89,8 +110,17 @@ public abstract class BaseRedisTest implements ContextBasedTest {
         if (nonNull(afterTransactionAssertion)) {
             final RedisConnection connection = connectionFactory.getConnection();
             afterTransactionAssertion.assertAfterTransaction(connection);
+            connection.close();
         }
     }
+
+//    @
+//    public void destroy() {
+//        redisServer.stop();
+////        if (Objects.nonNull(emailEmbeddedRedis)) {
+////            emailEmbeddedRedis.stopRedis();
+////        }
+//    }
 
     @FunctionalInterface
     public interface BeforeTransactionAssertion {
@@ -105,6 +135,11 @@ public abstract class BaseRedisTest implements ContextBasedTest {
     }
 
 
+    @PreDestroy
+    public void redisServerDestroy() {
+        System.exit(-1);
+    }
+
     @Configuration
     @PropertySource("classpath:redis-test.properties")
     //@TestPropertySource("classpath:redis-test.yml")
@@ -116,6 +151,7 @@ public abstract class BaseRedisTest implements ContextBasedTest {
         private JedisConnectionFactory connectionFactory;
 
         public JedisContextConfiguration() throws IOException {
+            System.out.println("AAAAH");
             int redisPort = randomFreePort();
 
 //            emailEmbeddedRedis = new EmailEmbeddedRedis(redisPort).startRedis();
@@ -128,19 +164,18 @@ public abstract class BaseRedisTest implements ContextBasedTest {
             JedisShardInfo shardInfo = new JedisShardInfo("localhost", redisPort);
             connectionFactory = new JedisConnectionFactory();
             connectionFactory.setShardInfo(shardInfo);
-        }
-
-        @PreDestroy
-        void destroy() {
-            redisServer.stop();
-            if (Objects.nonNull(emailEmbeddedRedis)) {
-                emailEmbeddedRedis.stopRedis();
-            }
+            connectionFactory.setUsePool(true);
+            connectionFactory.getPoolConfig().setMaxTotal(10_000);
         }
 
         @Bean
         public PropertySourcesPlaceholderConfigurer propertyConfigInDev() {
             return new PropertySourcesPlaceholderConfigurer();
+        }
+
+        @Bean
+        public RedisServer redisServer() {
+            return redisServer;
         }
 
         @Bean
@@ -171,15 +206,40 @@ public abstract class BaseRedisTest implements ContextBasedTest {
             return template;
         }
 
-        @Bean
-        public PlatformTransactionManager transactionManager() throws SQLException {
-            return new DataSourceTransactionManager(dataSource());
-        }
+//        @Bean
+//        public PlatformTransactionManager transactionManager() throws SQLException {
+//            return new DataSourceTransactionManager(dataSource);
+//
+////            return new AbstractPlatformTransactionManager() {
+////                @Override
+////                protected Object doGetTransaction() throws TransactionException {
+////                    return null;
+////                }
+////
+////                @Override
+////                protected void doBegin(Object o, TransactionDefinition transactionDefinition) throws TransactionException {
+////
+////                }
+////
+////                @Override
+////                protected void doCommit(DefaultTransactionStatus defaultTransactionStatus) throws TransactionException {
+////
+////                }
+////
+////                @Override
+////                protected void doRollback(DefaultTransactionStatus defaultTransactionStatus) throws TransactionException {
+////
+////                }
+////            };
+//        }
+//
+//        DataSource dataSource = mock(DataSource.class);
+
 
         @Bean
         public DataSource dataSource() throws SQLException {
             DataSource dataSource = mock(DataSource.class);
-            Mockito.when(dataSource.getConnection()).thenReturn(mock(Connection.class));
+//            when(dataSource.getConnection()).thenReturn(mock(Connection.class));
             return dataSource;
         }
 
@@ -193,5 +253,19 @@ public abstract class BaseRedisTest implements ContextBasedTest {
         }
 
     }
+
+
+
+    public void afterTestClass(TestContext testContext){
+            redisServer.stop();
+//            System.out.println("NOOOOOOOO");
+//            if (nonNull(emailEmbeddedRedis)) {
+//                emailEmbeddedRedis.stopRedis();
+//            }
+    }
+    public void	afterTestMethod(TestContext testContext){}
+    public void	beforeTestClass(TestContext testContext){}
+    public void	beforeTestMethod(TestContext testContext){}
+    public void	prepareTestInstance(TestContext testContext){}
 
 }