+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/templating/mail/service/defaultimpl/RedisBasedPersistenceServiceConstantsTest.java
@@ -12,38 +12,62 @@ import static org.junit.Assert.*;
 
 public class RedisBasedPersistenceServiceConstantsTest {
 
+    private static final String EXPECTED_ORDERING_KEY_PREFIX = "priority-level:";
+
     @Rule
     public ExpectedException expectedException = ExpectedException.none();
 
     @Rule
-    public Timeout timeout = new Timeout(120, TimeUnit.SECONDS);
+    public Timeout timeout = new Timeout(10, TimeUnit.SECONDS);
 
     @Rule
     public final JUnitSoftAssertions assertions = new JUnitSoftAssertions();
 
     @Test
-    public void orderingKey() throws Exception {
-
-    }
-
-    @Test
-    public void orderingKeyPrefix() throws Exception {
-
-    }
-
-    @Test
-    public void whenOrderingKeyPrefixThenShouldNotChange() {
+    public void shouldThrowExceptionWhenOrderingKeyIsConstructedFromNegativeNumber() throws Exception {
         //Arrange
-        final String expectedPrefix = "priority-level:";
+        expectedException.expect(IllegalArgumentException.class);
+        expectedException.expectMessage("Priority level must be a positive integer number");
 
+        //Act
+        RedisBasedPersistenceServiceConstants.orderingKey(-1);
+    }
+
+    @Test
+    public void shouldCreateOrderingKeyPrefixFromZero() throws Exception {
+        //Arrange
+        int priorityLevel = 0;
+        String expectedOrderingKey = EXPECTED_ORDERING_KEY_PREFIX + priorityLevel;
+
+        //Act
+        String givenOrderingKey = RedisBasedPersistenceServiceConstants.orderingKey(priorityLevel);
+
+        //Assert
+        assertions.assertThat(givenOrderingKey).isEqualTo(expectedOrderingKey);
+    }
+
+    @Test
+    public void shouldCreateOrderingKeyPrefixFromPositiveNumber() throws Exception {
+        //Arrange
+        int priorityLevel = 10;
+        String expectedOrderingKey = EXPECTED_ORDERING_KEY_PREFIX + priorityLevel;
+
+        //Act
+        String givenOrderingKey = RedisBasedPersistenceServiceConstants.orderingKey(priorityLevel);
+
+        //Assert
+        assertions.assertThat(givenOrderingKey).isEqualTo(expectedOrderingKey);
+    }
+
+    @Test
+    public void shouldOrderingKeyPrefixNotChange() {
         //Act
         final String givenPrefix = RedisBasedPersistenceServiceConstants.orderingKeyPrefix();
 
         //Assert
         assertions.assertThat(givenPrefix)
-                .describedAs("Changing constants should be avoided")
-                .isEqualTo(expectedPrefix);
+                .describedAs("Changing constants should be worth prison for life")
+                .isEqualTo(EXPECTED_ORDERING_KEY_PREFIX);
     }
-
 
 }